# src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dupip Morpheus API
  version: 1.0.0
  description: >
    HTTP API for the Dupip application. Authentication uses Clerk session
    cookies; most routes require an authenticated user. Errors use the shape
    `{ "error": "message" }`.
servers:
  - url: https://www.dupip.com
    description: Production
  - url: https://beta.dupip.com
    description: Nightly
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/debug/task-state:
    get:
      tags: [tasks]
      summary: Debug task/job completion state
      parameters:
        - { name: taskId, in: query, required: true, schema: { type: string } }
        - { name: date, in: query, schema: { type: string } }
      responses:
        '200': { description: Task state }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __session